Simplify database setup code in preparation for introducing validation
Acquiring the mongo connection is now done on startup so it can fail
fast and be ready before the first client request.

We don't have any reasonable expectation of it failing though anyway, as we're using an in-memory server with a (now) dynamically allocated port. For this reason the try-catch around the mongo connect call is removed.